refactor: split main routes
